Authentication with FNC Connect

+++ eventPublisher.data/Entities/ApplicationEntity.cs
@@ -6,7 +6,7 @@ namespace eventPublisher.data.entities
     public class ApplicationEntity
     {
         [Key]
-        public int ApplicationId { get; set; }
+        public long ApplicationId { get; set; }
         public string Name { get; set; }
         public DateTime InsertedUtc { get; set; }
     }
