+++ eventPublisher.data/Migrations/EventPublisherContextModelSnapshot.cs
@@ -21,7 +21,7 @@ namespace eventPublisher.data.Migrations
 
             modelBuilder.Entity("eventPublisher.data.entities.ApplicationEntity", b =>
                 {
-                    b.Property<int>("ApplicationId")
+                    b.Property<long>("ApplicationId")
                         .ValueGeneratedOnAdd()
                         .HasColumnName("application_id");
 
@@ -45,7 +45,7 @@ namespace eventPublisher.data.Migrations
                         .ValueGeneratedOnAdd()
                         .HasColumnName("event_id");
 
-                    b.Property<int>("ApplicationId")
+                    b.Property<long>("ApplicationId")
                         .HasColumnName("application_id");
 
                     b.Property<DateTime>("InsertedUtc")
